Refactored runtime code
Try to simplify existing code differences between various runtime/OS
variants.
Create a separate module for each runtime.
